==> notebooks/2_feature_selection.ipynb <==
import pandas as pd
import numpy as np

<br>
<br>
<br>

### Data Collection

# loading train data
X = pd.read_csv("../Training/X_train.csv")
X.head(3)

# loading test data
y = pd.read_csv("../Training/y_train.csv")
y.head(3)

# combining together
total = pd.concat([X, y], axis=1)
total.head()

<br>
<br>
<br>

### Train & Test Data

Before we do feature selection

Seperating given data into training data and testing data<br>
All the further steps will be carried out on training data<br>
Testing data will not be touched until a model is ready to be evaluated<br>
This will help in avoding data leakage and achieve unbiased testing of the model

from sklearn.model_selection import train_test_split

# test_size 0f 5,000
train_data, test_data = train_test_split(total, test_size=5000, stratify=total['Dependent_Variable'], shuffle=True)

print(train_data.shape)
print(test_data.shape)

# saving train data
train_data.to_csv("../pipeline_data/train_data.csv", index=None)

# saving test data
test_data.to_csv("../pipeline_data/test_data.csv", index=None)

<br>
<br>
<br>

### Feature Selection

# these feature are taken from EDA
informational = ['Unique_ID']
high_missing_values = ['N32', 'N27', 'N26', 'N25', 'N31', 'N30', 'N29', 'N28']
high_correlation = ['N6', 'N5', 'N20', 'N34']

features_to_drop = informational + high_missing_values + high_correlation

train_useful = train_data.drop(features_to_drop, axis=1)

train_useful.head()

train_useful.shape

<br>
<br>
<br>

### Saving

# saving feature selected dataset
train_useful.to_csv("../pipeline_data/2_train_feature_selected.csv", index=None)
