==> notebooks/3_data_preparation.ipynb <==
import pandas as pd
import numpy as np

<br>
<br>
<br>

### Data Collection

# loading feature selected data
train_data = pd.read_csv("../pipeline_data/2_train_feature_selected.csv")
train_data.head(3)

# making a copy
train_data_updated = train_data.copy()

<br>
<br>
<br>

### Data Preparation

#### Categorical

Reducing cardinality of high cardinal features by keeping<br>
top n values and replacing other by 999<br>
<br>
The value for n is decided by inspecting the feature distribution 

cat_features = ['C1', 'C2', 'C3', 'C4', 'C5', 'C7']

# no. of unique values
train_data_updated[cat_features].apply(lambda x: x.nunique())

train_data_updated.C1.value_counts()

Keeping all 3 values

<br>

train_data_updated.C2.value_counts() / len(train_data_updated) * 100

c2_values_to_keep = train_data_updated.C2.value_counts()[:7].index
c2_values_to_keep

train_data_updated.C2 = train_data_updated.C2.apply(lambda x: x if x in c2_values_to_keep else 999)
train_data_updated[cat_features].head(3)

<br>

train_data_updated.C3.value_counts() / len(train_data_updated) * 100

c3_values_to_keep = train_data_updated.C3.value_counts()[:7].index
c3_values_to_keep

train_data_updated.C3 = train_data_updated.C3.apply(lambda x: x if x in c3_values_to_keep else 999)
train_data_updated[cat_features].head(3)

<br>

(train_data_updated.C4.value_counts() / len(train_data_updated) * 100).head(15)

c4_values_to_keep = train_data_updated.C4.value_counts()[:5].index
c4_values_to_keep

train_data_updated.C4 = train_data_updated.C4.apply(lambda x: x if x in c4_values_to_keep else 999)
train_data_updated[cat_features].head(3)

<br>

train_data_updated.C5.value_counts() / len(train_data_updated) * 100

c5_values_to_keep = train_data_updated.C5.value_counts()[:4].index
c5_values_to_keep

train_data_updated.C5 = train_data_updated.C5.apply(lambda x: x if x in c5_values_to_keep else 999)
train_data_updated[cat_features].head(3)

<br>

train_data_updated.C7.value_counts() / len(train_data_updated) * 100

c7_values_to_keep = train_data_updated.C7.value_counts()[:5].index
c7_values_to_keep

train_data_updated.C7 = train_data_updated.C7.apply(lambda x: x if x in c7_values_to_keep else 999)
train_data_updated[cat_features].head(3)

<br>

# updated unique value counts
train_data_updated[cat_features].apply(lambda x: x.nunique())

<br>
<br>
<br>

#### Numerical Data

numerical_features = train_data_updated.columns[8:-1]
numerical_features

# outlier limit for each feature
# derived from EDA

outlier_limit = {
    'N1': 45,
    'N2': 530,
    'N3': 4.3,
    'N4': 40,
    'N7': 95,
    'N8': 33,
    'N9': 45,
    'N10': 4500,
    'N10.1': 32,
    'N11': 80,
    'N12': 100_000,
    'N14': 100,
    'N15': 13,
    'N16': 2,
    'N17': 450_000,
    'N18': 1.8,
    'N19': 260_000,
    'N21': 1,
    'N22': 10,
    'N23': 2100,
    'N24': 50_000,
    'N33': 500,
    'N35': 54,
}

Instead of caping or removing the outliers, we'll replace them with null values.<br>
So they can be imputed later. This will prevent data loss.

# replacing outliers with na

for feature in numerical_features:
    train_data_updated[feature] = train_data_updated[feature].apply(lambda x: x if x <= outlier_limit[feature] else np.nan)

<br>
<br>

# missing values in each feature

train_data_updated[numerical_features].isna().sum()

All the instances where N19 is null, respective N2, N4, N12 and N16 - N22 are also null<br>
This indicates MNAR `Missing Not At Random`

We can remove these datapoints, but imputing them provides better result<br>
Also it avoids loss of data

<br>

Imputing missing values using Aggregate (median)

for feature in numerical_features:
    train_data_updated[feature].fillna(train_data_updated[feature].median(), inplace=True)

`OR`

Imputing missing values using KNNImputer

from sklearn.impute import KNNImputer

imputer = KNNImputer(n_neighbors=5)

# imputing & updating dataframe

imputed_data = imputer.fit_transform(train_data_updated[numerical_features])
train_data_updated[numerical_features] = pd.DataFrame(imputed_data, columns=train_data_updated[numerical_features].columns, index=train_data_updated[numerical_features].index)
train_data_updated[numerical_features].head()

<br>
<br>

# checking for null values
train_data_updated.isna().any().any()

<br>
<br>

Converting skewed distribution using `log_transformation` to have normal-like distribution

# features derived from EDA

skewed_features = [
    'N2',
    'N9',
    'N17',
    'N19',
]

# applying log(1+x)

for feature in skewed_features:
    train_data_updated[feature] = np.log1p(train_data_updated[feature])

<br>
<br>
<br>

# updated train data
train_data_updated.head()

# saving prepared dataset
train_data_updated.to_csv("../pipeline_data/3_train_prepared.csv", index=None)
