==> notebooks/5_data_preprocessing.ipynb <==
import pandas as pd
import numpy as np

<br>
<br>
<br>

### Data Collection

# loading prepared data
train_data = pd.read_csv("../pipeline_data/3_train_prepared.csv")
train_data.head()

# checking for null values
train_data.isna().any().any()

<br>
<br>
<br>

### Data Proprocessing

from sklearn.preprocessing import RobustScaler

#### Boolean data

bool_data = train_data[['C6', 'C8']].copy()
bool_data.head()

# replacing
# True -> 1
# False -> 0
bool_data_encoded = bool_data.applymap(lambda x: 1 if x else 0)
bool_data_encoded.head()

bool_data_encoded.shape

<br>
<br>

#### Categorical data

cat_data = train_data[['C1', 'C2', 'C3', 'C4', 'C5', 'C7']].copy()
cat_data.head()

# changing dtype to string
cat_data = cat_data.astype('string')
cat_data.head()

<br>

Here, we are using `pd.get_dummies` instead of OneHotEncoder because of 2 reasons:
- OHE returns a sparse matrix
- This matrix does not contain column names

We need column names because when we will do Recursive Feature Elimination,<br>
we can easily see which columns have low feature importance.<br>

This will make removing those features easy and it also provides better interpretability.

cat_data_encoded = pd.get_dummies(cat_data)
cat_data_encoded.head()

cat_data_encoded.shape

<br>
<br>

#### Numerical data

As the data is skewed, using robust scaler provides<br>
better accuracy than standard or min-max scaler

scaler = RobustScaler()

num_data = train_data.iloc[:, 8:-1].copy()
num_data.head()

num_data_encoded = scaler.fit_transform(num_data)
# converting ndarray to dataframe
num_data_encoded = pd.DataFrame(data=num_data_encoded, columns=num_data.columns, index=num_data.index)
num_data_encoded.head()

num_data_encoded.shape

<br>
<br>

total_data = pd.concat([bool_data_encoded, cat_data_encoded, num_data_encoded, train_data[['Dependent_Variable']]], axis=1)
total_data.head()

total_data.columns

total_data.isna().any().any()

# saving preprocessed data
total_data.to_csv("../pipeline_data/5_train_preprocessed.csv", index=None)
