==> notebooks/7_data_pipeline_tuning.ipynb <==
This notebook optimize the data pipeline parameters

Parameters:
- Remove MNAR data or impute it
- How to handle skewed numerical features
- Missing values imputation technique (Aggregate or KNN)
- Scaler for numerical data (Standard, MinMax or Robust)

Selected:
- Impute MNAR data
- Log transformation
- KNN has slightly better accuracy but Aggregate is much faster
- Robust scaler

<br>
<br>
<br>

import pandas as pd
import numpy as np

from pathlib import Path
import sys
sys.path.insert(1, str(Path("../").resolve()))

# importing data pipelines
from scripts.pipelines.a_feature_selection import FeatureSelectionPipeline
from scripts.pipelines.b_data_preparation import DataPreparationPipeline
from scripts.pipelines.d_data_preprocessing import DataPreprocessingPipeline
from scripts.pipelines.e_feature_selection_2 import FeatureSelectionPipeline2
from scripts.utils import load_csv

<br>
<br>
<br>

# loading train data
train_data = load_csv("../pipeline_data/1_train_data.csv")
train_data.head(3)

train_data.shape

<br>
<br>
<br>

from sklearn.pipeline import Pipeline

# creating full pipeline
full_pipeline = Pipeline([
    ('feature_selection', FeatureSelectionPipeline()),
    ('data_preparation', DataPreparationPipeline(imputation='aggr')),
    ('data_preprocessing', DataPreprocessingPipeline()),
    ('feature_selection_2', FeatureSelectionPipeline2()),
])

# transforming train data
train_transformed = full_pipeline.fit_transform(train_data)
train_transformed.head(3)

train_transformed.shape

train_transformed.isna().any().any()

<br>
<br>
<br>

from sklearn.model_selection import cross_val_score
from sklearn.metrics import roc_auc_score

from lightgbm import LGBMClassifier

<br>

best_params = {
    'n_estimators': 400,
    'learning_rate': 0.10212499331506983,
    'num_leaves': 1160,
    'max_depth': 8,
    'min_data_in_leaf': 200,
    'lambda_l1': 0,
    'lambda_l2': 15,
    'min_gain_to_split': 7.151260846950484,
    'bagging_fraction': 0.7,
    'bagging_freq': 1,
    'feature_fraction': 0.5
}

# seperating features and target
model_x = train_transformed.iloc[:, :-1]
model_y = np.ravel(train_transformed.iloc[:, -1])

clf = LGBMClassifier(**best_params)

clf.fit(model_x, model_y)

# score on training data
predictions = clf.predict_proba(model_x)[:, 1]
roc_auc_score(model_y, predictions)

# cross validation score
scores = cross_val_score(clf, X=model_x, y=model_y, scoring='roc_auc', cv=10, n_jobs=-1)
scores.mean()

# feature importance
pd.Series(clf.feature_importances_.round(3), index=model_x.columns).sort_values(ascending=False)[:10]
