==> notebooks/8_model_selection.ipynb <==
import pandas as pd
import numpy as np

import warnings
warnings.filterwarnings('ignore')

<br>
<br>
<br>

### Data Collection

# loading preprocessed & feature selected data
train_data = pd.read_csv("../pipeline_data/6_train_feature_selected.csv")
train_data.head(3)

train_data.shape

# seperating features
X = train_data.iloc[:, :-1]
X.head()

# seperating target
y = train_data.iloc[:, -1]
y = np.ravel(y)
print(y.shape)
y

<br>
<br>
<br>

### LightGBM

Hyperparameter Tuning

import optuna
from lightgbm import LGBMClassifier
from optuna.integration import LightGBMPruningCallback
from sklearn.metrics import roc_auc_score
from sklearn.model_selection import StratifiedKFold, cross_val_score

# optuna objective function
def objective(trial, X, y):

    # parameters grid
    param_grid = {
        "n_estimators": trial.suggest_int("n_estimators", 100, 2000, step=100),
        "learning_rate": trial.suggest_float("learning_rate", 0.01, 0.3),
        "num_leaves": trial.suggest_int("num_leaves", 20, 3000, step=20),
        "max_depth": trial.suggest_int("max_depth", 3, 12),
        "min_data_in_leaf": trial.suggest_int("min_data_in_leaf", 200, 10000, step=100),
        "lambda_l1": trial.suggest_int("lambda_l1", 0, 100, step=5),
        "lambda_l2": trial.suggest_int("lambda_l2", 0, 100, step=5),
        "min_gain_to_split": trial.suggest_float("min_gain_to_split", 0, 15),
        "bagging_fraction": trial.suggest_float(
            "bagging_fraction", 0.2, 0.95, step=0.1
        ),
        "bagging_freq": trial.suggest_categorical("bagging_freq", [1]),
        "feature_fraction": trial.suggest_float(
            "feature_fraction", 0.2, 0.95, step=0.1
        )
    }

    # 5 fold cv
    cv = StratifiedKFold(n_splits=5, shuffle=True)
    # will hold scores
    cv_scores = np.empty(5)

    for idx, (train_idx, test_idx) in enumerate(cv.split(X, y)):
        X_train, X_test = X.iloc[train_idx], X.iloc[test_idx]
        y_train, y_test = y[train_idx], y[test_idx]

        # training model
        model = LGBMClassifier(objective="binary", verbose=-1, **param_grid)
        model.fit(
            X_train,
            y_train,
            eval_set=[(X_test, y_test)],
            eval_metric="auc",
            early_stopping_rounds=100,
            callbacks=[
                LightGBMPruningCallback(trial, "auc")
            ]
        )

        # getting auc score
        preds = model.predict_proba(X_test)[:, 1]
        cv_scores[idx] = roc_auc_score(y_test, preds)

    # return mean score
    return np.mean(cv_scores)

# optuna study
study = optuna.create_study(direction="maximize", study_name="LGBM Classifier")
func = lambda trial: objective(trial, X, y)
# optimizing the study
study.optimize(func, timeout=180)

# [I 2023-03-24 12:32:15,524] Trial 321 finished with value: 0.7520746168126591 and parameters: {'n_estimators': 700, 'learning_rate': 0.1921286133667085, 'num_leaves': 2360, 'max_depth': 9, 'min_data_in_leaf': 1000, 'lambda_l1': 30, 'lambda_l2': 20, 'min_gain_to_split': 2.993720055467331, 'bagging_fraction': 0.9, 'bagging_freq': 1, 'feature_fraction': 0.5}. Best is trial 233 with value: 0.756908868345348.

# best set of parameters
best_params = {
    'n_estimators': 700,
    'learning_rate': 0.1921286133667085,
    'num_leaves': 2360,
    'max_depth': 9,
    'min_data_in_leaf': 1000,
    'lambda_l1': 30,
    'lambda_l2': 20,
    'min_gain_to_split': 2.993720055467331,
    'bagging_fraction': 0.9,
    'bagging_freq': 1,
    'feature_fraction': 0.5
    }

<br>
<br>
<br>

Model performance with best parameters

# classifier with best parameters
clf = LGBMClassifier(**best_params)

clf.fit(X, y)

# training score
predictions = clf.predict_proba(X)[:, 1]
roc_auc_score(y, predictions)

# cross val score
scores = cross_val_score(clf, X=X, y=y, scoring='roc_auc', cv=10, n_jobs=-1)
scores.mean()

# feature importance
pd.Series(clf.feature_importances_, index=X.columns).sort_values(ascending=False)[:10]
