==> notebooks/10_making_prediction.ipynb <==
import pandas as pd
import numpy as np

from pathlib import Path
import sys
sys.path.insert(1, str(Path("../").resolve()))

# importing data pipelines
from scripts.pipelines.a_feature_selection import FeatureSelectionPipeline
from scripts.pipelines.b_data_preparation import DataPreparationPipeline
from scripts.pipelines.d_data_preprocessing import DataPreprocessingPipeline
from scripts.pipelines.e_feature_selection_2 import FeatureSelectionPipeline2
from scripts.utils import load_csv

<br>
<br>
<br>

### Training on full data

Now that we have our model ready, let's train it on complete dataset one last time.

# loading raw x_train
x_train = pd.read_csv("../Training/X_train.csv")
x_train.head(3)

# loading raw y_train
y_train = pd.read_csv("../Training/Y_train.csv")
y_train.head(3)

# combining training data
total_train = pd.concat([x_train, y_train], axis=1)
total_train.head(3)

# loading raw x_test
x_test = pd.read_csv("../Test/X_test.csv")
x_test.head(3)

print(total_train.shape)
print(x_test.shape)

<br>
<br>
<br>

### Data Pipeline

from sklearn.pipeline import Pipeline

# full data pipeline
full_pipeline = Pipeline([
    ('feature_selection', FeatureSelectionPipeline()),
    ('data_preparation', DataPreparationPipeline(imputation='aggr')),
    ('data_preprocessing', DataPreprocessingPipeline()),
    ('feature_selection_2', FeatureSelectionPipeline2()),
])

# transforming training data
train_transformed = full_pipeline.fit_transform(total_train)
train_transformed.head()

print(train_transformed.shape)
train_transformed.isna().any().any()

<br>
<br>

# transforming test data
x_test_transformed = full_pipeline.transform(x_test)
x_test_transformed.head(3)

print(x_test_transformed.shape)
x_test_transformed.isna().any().any()

# verifying columns
pd.Series(train_transformed.columns[:-1] == x_test_transformed.columns).value_counts()

<br>
<br>
<br>

from lightgbm import LGBMClassifier
from sklearn.model_selection import cross_val_score
from sklearn.metrics import roc_auc_score

Training LGBM model

# x_train & y_train
x_train = train_transformed.drop(['Dependent_Variable'], axis=1)
y_train = train_transformed[['Dependent_Variable']]
y_train = np.ravel(y_train)
print(x_train.shape)
print(y_train.shape)

# best set of parameters
best_params = {
    'n_estimators': 700,
    'learning_rate': 0.1921286133667085,
    'num_leaves': 2360,
    'max_depth': 9,
    'min_data_in_leaf': 1000,
    'lambda_l1': 30,
    'lambda_l2': 20,
    'min_gain_to_split': 2.993720055467331,
    'bagging_fraction': 0.9,
    'bagging_freq': 1,
    'feature_fraction': 0.5
}

# model
clf = LGBMClassifier(**best_params, random_state=42)

# training our model
clf.fit(x_train, y_train)

# score on training data
predictions = clf.predict_proba(x_train)[:, 1]
roc_auc_score(y_train, predictions)

# cross validation score
scores = cross_val_score(clf, X=x_train, y=y_train, scoring='roc_auc', cv=10, n_jobs=-1)
scores.mean()

<br>
<br>

Making predictions on test data

predictions = clf.predict_proba(x_test_transformed)[:, 1]

predictions

<br>
<br>
<br>

### Submission

# loading sample submission
submission = pd.read_csv("../sample_submission_file.csv")
submission.head()

# replacing sample values with actual predictions
submission['Class_1_Probability'] = predictions
submission.head()

# saving as a csv file
submission.to_csv("../final_submission.csv", index=None)
